Check existence of requirements.txt before calling. Closes #21

+++ scripts/build-env.sh
@@ -27,7 +27,9 @@ function buildEnvironment(){
 
     export PYTHONPATH=${PYTHONPATH}:${SRCROOT}
 
-    ${PYTHONENV}/pip install -r ${REQUIREMENTS}
+    if [ -f ${REQUIREMENTS} ]; then
+        ${PYTHONENV}/pip install -r ${REQUIREMENTS}
+    fi
 }
 
 buildEnvironment
